Removed 'ReadEntry' M2M relation between Subscriptions and Entries and added 'Visited' M2M relation between Users and Entries.
This way, a user's data about feed entries remain even if he cancels his
subscription. This is useful if he later re-adds it or if you want to
display it elsewhere. Also, it should be a faster lookup, you don't have
to go over the Subscription table to ask for read entries for a given
user.